mcp-server-youtube/pyproject.toml
[project]
name = "mcp-server-youtube"
version = "0.1.1"
description = "MCP Server for searching YouTube videos and retrieving transcripts with enhanced transcript handling"
authors = [
    { name = "XyAgent", email = "agent@xyber.inc" },
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "mcp>=1.8.0,<2.0.0",
    "pydantic-settings>=2.0",
    "uvicorn[standard]",
    "fastmcp>=2.3.4",
    "fastapi>=0.115.12",
    # YouTube-specific dependencies
    "youtube-transcript-api>=0.6.1",
    "youtube-search-python>=1.6.6",
    "google-api-python-client>=2.0.0",
    # Additional dependencies for transcript handling
    "xmltodict>=0.13.0",
]

[dependency-groups]
dev = [
    "pytest",
    "httpx",
    "ruff",
    "black",
    "isort",
]
debug = [
    "debugpy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_youtube"]
